+++ rs/ic_os/dev_test_tools/setupos-inject-configuration/src/main.rs
@@ -42,6 +42,9 @@ struct Cli {
 #[derive(Args)]
 struct ConfigIni {
     #[arg(long)]
+    node_reward_type: Option<String>,
+
+    #[arg(long)]
     ipv6_prefix: Option<String>,
 
     #[arg(long)]
@@ -209,6 +212,7 @@ async fn write_config(path: &Path, cfg: &ConfigIni) -> Result<(), Error> {
     let mut f = File::create(path).context("failed to create config file")?;
 
     let ConfigIni {
+        node_reward_type,
         ipv6_prefix,
         ipv6_gateway,
         ipv4_address,
@@ -217,6 +221,10 @@ async fn write_config(path: &Path, cfg: &ConfigIni) -> Result<(), Error> {
         domain,
         verbose,
     } = cfg;
+
+    if let Some(node_reward_type) = node_reward_type {
+        writeln!(&mut f, "node_reward_type={}", node_reward_type)?;
+    }
 
     if let (Some(ipv6_prefix), Some(ipv6_gateway)) = (ipv6_prefix, ipv6_gateway) {
         // Always write 4 segments, even if our prefix is less.